Extract some classes
